+Валидаторы

+++ src/main/java/validators/primitive/ValidateInt.java
@@ -2,7 +2,7 @@ package validators.primitive;
 
 public class ValidateInt
 {
-    public boolean moreZero(int i)
+    public boolean moreZero(long i)
     {
         return i>0;
     }
